refactor: rename App to routes and clean it
